packages/agents/requirements.txt
# Core dependencies
pydantic>=2.5.0
python-dotenv>=1.0.0

# Message bus
pika>=1.3.2
protobuf>=4.25.0
grpcio-tools>=1.60.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# NLP and ML
spacy>=3.7.0
transformers>=4.35.0
torch>=2.1.0

# LLM integration
openai>=1.3.0
langchain>=0.1.0

# Database clients
influxdb-client>=1.38.0
pymongo>=4.6.0
redis>=5.0.0

# HTTP clients
httpx>=0.25.0
aiohttp>=3.9.0

# Redis async client
redis[hiredis]>=5.0.0

# Scheduling
apscheduler>=3.10.0

# OpenTelemetry tracing
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-instrumentation-pymongo>=0.42b0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.12.0
isort>=5.13.0
pylint>=3.0.0
mypy>=1.7.0
